Fixes default values on dead forms
Forms submitted outside of LiveView submit empty strings as params for
empty fields. Empty fields are problematic for default values because
"" is a value, and the resource won't set the default.

To allow the dead view forms, I have added a new
form option. `empty_fields_unset` takes a list of field names. The form
will remove empty fields in the `empty_fields_unset` from the params map
before the params are processed.

The type in the documentation is recursive at the tuple value level so
I stoped and made it be `:any` at that point.

closes #52